refactor: 공통 필드 묶고 팀 생성과 조회 함수를 하나의 파일로 통합
